PetFood migration error fixed

+++ db/migrate/20210317111857_create_pet_foods.rb
@@ -2,7 +2,7 @@ class CreatePetFoods < ActiveRecord::Migration[6.1]
   def change
     create_table :pet_foods do |t|
       t.string :name
-      t.double :price
+      t.float :price
 
       t.timestamps
     end
